requirements.txt
Flask==2.3.3
gunicorn==21.2.0
flask-cors==4.0.0
requests==2.31.0
anthropic==0.3.11
python-dotenv==1.0.0
line-bot-sdk==3.2.0